+++ core/src/main/java/in/testpress/database/entities/UpcomingContentRemoteKeys.kt
@@ -1,0 +1,12 @@
+package `in`.testpress.database.entities
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class UpcomingContentRemoteKeys(
+    @PrimaryKey val contentId: Long,
+    val prevKey: Int?,
+    val nextKey: Int?,
+    val courseId: Long
+)
